new functions and new trigger
Functions now run based on the selection of a checkbox in the sheet itself.
setRanks function runs down the list of ranked countries and re-writes their order in ascending numbers up to 20.
orderRank function orders the data range of the sheet in ascending order by means of rank number.
orderActive function orders the data range of the sheet by color, so redundant currencies and countries falling out of the viable 10 get pushed to the bottom.